[refac] move ApiResponse.java, BaseResponse.java
