+++ src/main/java/org/hankki/hankkiserver/api/auth/controller/AuthController.java
@@ -3,8 +3,8 @@ package org.hankki.hankkiserver.api.auth.controller;
 import jakarta.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
 import org.hankki.hankkiserver.auth.UserId;
-import org.hankki.hankkiserver.common.ApiResponse;
-import org.hankki.hankkiserver.common.BaseResponse;
+import org.hankki.hankkiserver.common.response.ApiResponse;
+import org.hankki.hankkiserver.common.response.BaseResponse;
 import org.hankki.hankkiserver.common.code.SuccessCode;
 import org.hankki.hankkiserver.api.auth.service.AuthService;
 import org.hankki.hankkiserver.api.auth.controller.request.UserLoginRequest;
