+++ src/main/java/org/hankki/hankkiserver/auth/filter/ExceptionHandlerFilter.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.hankki.hankkiserver.common.BaseResponse;
+import org.hankki.hankkiserver.common.response.BaseResponse;
 import org.hankki.hankkiserver.common.code.ErrorCode;
 import org.hankki.hankkiserver.common.exception.UnauthorizedException;
 import org.springframework.http.HttpStatus;
